+++ apps/build.sh
@@ -1,3 +1,0 @@
-cd sparkoin
-sbt assembly
-cd ..