+++ build.sh
@@ -12,7 +12,3 @@ then
 else
   sh build-one.sh "$1"
 fi
-cd -
-cd apps
-sh build.sh
-cd -