+++ start.sh
@@ -4,4 +4,5 @@ then
      docker-machine ssh sparkoin "sudo sh /var/lib/boot2docker/bootlocal.sh"
      eval $(docker-machine env sparkoin)
 fi
-docker-compose up --no-color "$@"
+docker-compose up --no-color -d
+./enter.sh sparkoin_jupyter "cd /app/apps/sparkoin ; sbt assembly"