+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.0.18
 
+* fixed hadoop in order to format datanode
 * fixed jupyter - can now load assemblies
 * added sparkoin library